Split content and back server to two files.
