+++ examples/config_backup_server.yml
@@ -1,12 +1,12 @@
 # Content server parameters
-remote_server: 'f1.test.kbb1.com' # Default is localhost
-backup_file_listening_port: 4444 # TCP/IP port to get backup files from.
-remote_content_port: 3333 # TCP/IP port to get content data from.
+content_server_hostname: 'localhost' #'f1.test.kbb1.com' # Default is nil
+content_server_files_port: 4444 # TCP/IP port to connect as client to receive files to backup from content server.
+content_server_data_port: 3333 # TCP/IP port to connect as client to get content data.
 
 # Local files and logs
 log_file_name: '~/.bbfs/log/backup_server.log' # General log
 default_monitoring_log_path: '~/.bbfs/log/backup_file_monitoring.log' # Log for file changes
-content_data_path: '~/.bbfs/var/backup.data' # File to store all local contents, i.e., backed up files.
+local_content_data_path: '~/.bbfs/var/backup.data' # File to store all local contents, i.e., backed up files.
 
 # Logging options
 log_write_to_console: true # Will print all log messages to standard output. Default is false.
