+++ examples/config_content_server.yml
@@ -1,12 +1,11 @@
 # Backup server parameters
-remote_server: 'f2.test.kbb1.com' # Default is localhost
-backup_file_listening_port: 4444 # TCP/IP port to send backup files to.
-remote_content_port: 3333 # TCP/IP port to send content data to.
+local_files_port: 4444 # TCP/IP port to start server to send files to backup.
+local_content_data_port: 3333 # TCP/IP port to start server to send content data upon request from client.
 
 # Local files and logs
 log_file_name: '~/.bbfs/log/content_server.log' # General log
 default_monitoring_log_path: '~/.bbfs/log/content_file_monitoring.log' # Log for file changes
-content_data_path: '~/.bbfs/var/content.data' # File to store all contents, i.e., state of content server.
+local_content_data_path: '~/.bbfs/var/content.data' # File to store all contents, i.e., state of content server.
 
 # Logging options
 log_write_to_console: true # Will print all log messages to standard output. Default is false.
